REF: move registry, Registry to dtypes.base (#34830)
